+++ lib/bubble_painter.dart
@@ -5,25 +5,46 @@ class BubblePainter extends CustomPainter {
   BubblePainter({
     required this.clipper,
     required Color color,
+    required Color borderColor,
+    required double borderWidth,
+    required this.borderUp,
     required this.elevation,
     required this.shadowColor,
-  }) : _paint = Paint()
+  })   : _fillPaint = Paint()
           ..color = color
-          ..style = PaintingStyle.fill;
+          ..style = PaintingStyle.fill,
+        _strokePaint = borderWidth == 0 || borderColor == Colors.transparent
+            ? null
+            : (Paint()
+              ..color = borderColor
+              ..strokeWidth = borderWidth
+              ..strokeCap = StrokeCap.round
+              ..strokeJoin = StrokeJoin.round
+              ..style = PaintingStyle.stroke);
 
   final CustomClipper<Path> clipper;
+  final bool borderUp;
   final double elevation;
   final Color shadowColor;
 
-  final Paint _paint;
+  final Paint _fillPaint;
+  final Paint? _strokePaint;
 
   @override
   void paint(Canvas canvas, Size size) {
+    final clip = clipper.getClip(size);
+
     if (elevation != 0.0) {
-      canvas.drawShadow(clipper.getClip(size), shadowColor, elevation, false);
+      canvas.drawShadow(clip, shadowColor, elevation, false);
     }
 
-    canvas.drawPath(clipper.getClip(size), _paint);
+    if (borderUp) canvas.drawPath(clip, _fillPaint);
+
+    if (_strokePaint != null) {
+      canvas.drawPath(clip, _strokePaint!);
+    }
+
+    if (!borderUp) canvas.drawPath(clip, _fillPaint);
   }
 
   @override